requirements.txt
apache-airflow-providers-postgres==5.4.0
apache-airflow-providers-amazon==7.4.1